feat(godot_parser_library): add support for parsing .tres files
This commit adds support for parsing .tres files, which are Godot's resource files. The parser is implemented in the `tscn_tres_parser` module and uses the same approach as the existing .tscn parser.

The `parse_tres_file` function takes a string containing the contents of a .tres file and returns a `TSCNFile` object. The `TSCNFile` object contains the header, external resources, sub-resources, and nodes of the .tres file.

The parser is tested with a simple .tres file and the output is printed to the console.